[#5 fix] 클래스 네이밍 변경
- `SerialDate`라는 이름의 추상화 수준이 올바르지 못하다
