+++ src/main/java/org/jfree/date/DayAndMonthRule.java
@@ -105,7 +105,7 @@ public class DayAndMonthRule extends AnnualDateRule {
     public void setDayOfMonth(final int dayOfMonth) {
 
         // check arguments...
-        if ((dayOfMonth < 1) || (dayOfMonth > SerialDate.LAST_DAY_OF_MONTH[this.month])) {
+        if ((dayOfMonth < 1) || (dayOfMonth > DayDate.LAST_DAY_OF_MONTH[this.month])) {
             throw new IllegalArgumentException(
                 "DayAndMonthRule(): dayOfMonth outside valid range.");
         }
@@ -136,7 +136,7 @@ public class DayAndMonthRule extends AnnualDateRule {
     public void setMonth(final int month) {
 
         // check arguments...
-        if (!SerialDate.isValidMonthCode(month)) {
+        if (!DayDate.isValidMonthCode(month)) {
             throw new IllegalArgumentException("DayAndMonthRule(): month code not valid.");
         }
 
@@ -152,7 +152,7 @@ public class DayAndMonthRule extends AnnualDateRule {
      *
      * @return the date generated by this rule for the specified year (null permitted).
      */
-    public SerialDate getDate(final int yyyy) {
-        return SerialDate.createInstance(this.dayOfMonth, this.month, yyyy);
+    public DayDate getDate(final int yyyy) {
+        return DayDate.createInstance(this.dayOfMonth, this.month, yyyy);
     }
 }
