+++ src/main/java/org/jfree/date/DayOfWeekInMonthRule.java
@@ -65,7 +65,7 @@ public class DayOfWeekInMonthRule extends AnnualDateRule {
      * Default constructor: builds a rule for the first Monday in January by default.
      */
     public DayOfWeekInMonthRule() {
-        this(1, SerialDate.MONDAY, MonthConstants.JANUARY);
+        this(1, DayDate.MONDAY, MonthConstants.JANUARY);
     }
 
     /**
@@ -142,23 +142,23 @@ public class DayOfWeekInMonthRule extends AnnualDateRule {
      *
      * @return the date generated by the rule for the given year.
      */
-    public SerialDate getDate(final int year) {
-        SerialDate result;
-        if (this.count != SerialDate.LAST_WEEK_IN_MONTH) {
+    public DayDate getDate(final int year) {
+        DayDate result;
+        if (this.count != DayDate.LAST_WEEK_IN_MONTH) {
             // start at the beginning of the month
-            result = SerialDate.createInstance(1, this.month, year);
+            result = DayDate.createInstance(1, this.month, year);
             while (result.getDayOfWeek() != this.dayOfWeek) {
-                result = SerialDate.addDays(1, result);
+                result = DayDate.addDays(1, result);
             }
-            result = SerialDate.addDays(7 * (this.count - 1), result);
+            result = DayDate.addDays(7 * (this.count - 1), result);
 
         }
         else {
             // start at the end of the month and work backwards...
-            result = SerialDate.createInstance(1, this.month, year);
+            result = DayDate.createInstance(1, this.month, year);
             result = result.getEndOfCurrentMonth(result);
             while (result.getDayOfWeek() != this.dayOfWeek) {
-                result = SerialDate.addDays(-1, result);
+                result = DayDate.addDays(-1, result);
             }
 
         }
