+++ src/main/java/org/jfree/date/RelativeDayOfWeekRule.java
@@ -70,7 +70,7 @@ public class RelativeDayOfWeekRule extends AnnualDateRule {
      * Default constructor - builds a rule for the Monday following 1 January.
      */
     public RelativeDayOfWeekRule() {
-        this(new DayAndMonthRule(), SerialDate.MONDAY, SerialDate.FOLLOWING);
+        this(new DayAndMonthRule(), DayDate.MONDAY, DayDate.FOLLOWING);
     }
 
     /**
@@ -171,31 +171,31 @@ public class RelativeDayOfWeekRule extends AnnualDateRule {
      * @return The date generated by the rule for the given year (possibly 
      *         <code>null</code>).
      */
-    public SerialDate getDate(final int year) {
+    public DayDate getDate(final int year) {
 
         // check argument...
-        if ((year < SerialDate.MINIMUM_YEAR_SUPPORTED)
-            || (year > SerialDate.MAXIMUM_YEAR_SUPPORTED)) {
+        if ((year < DayDate.MINIMUM_YEAR_SUPPORTED)
+            || (year > DayDate.MAXIMUM_YEAR_SUPPORTED)) {
             throw new IllegalArgumentException(
                 "RelativeDayOfWeekRule.getDate(): year outside valid range.");
         }
 
         // calculate the date...
-        SerialDate result = null;
-        final SerialDate base = this.subrule.getDate(year);
+        DayDate result = null;
+        final DayDate base = this.subrule.getDate(year);
 
         if (base != null) {
             switch (this.relative) {
-                case(SerialDate.PRECEDING):
-                    result = SerialDate.getPreviousDayOfWeek(this.dayOfWeek, 
+                case(DayDate.PRECEDING):
+                    result = DayDate.getPreviousDayOfWeek(this.dayOfWeek,
                             base);
                     break;
-                case(SerialDate.NEAREST):
-                    result = SerialDate.getNearestDayOfWeek(this.dayOfWeek, 
+                case(DayDate.NEAREST):
+                    result = DayDate.getNearestDayOfWeek(this.dayOfWeek,
                             base);
                     break;
-                case(SerialDate.FOLLOWING):
-                    result = SerialDate.getFollowingDayOfWeek(this.dayOfWeek, 
+                case(DayDate.FOLLOWING):
+                    result = DayDate.getFollowingDayOfWeek(this.dayOfWeek,
                             base);
                     break;
                 default:
