+++ src/main/java/org/jfree/ui/SerialDateChooserPanel.java
@@ -61,7 +61,7 @@ import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.SwingConstants;
 
-import org.jfree.date.SerialDate;
+import org.jfree.date.DayDate;
 
 /**
  * A panel that allows the user to select a date.
@@ -79,7 +79,7 @@ public class SerialDateChooserPanel extends JPanel implements ActionListener {
     public static final Color DEFAULT_MONTH_BUTTON_COLOR = Color.lightGray;
 
     /** The date selected in the panel. */
-    private SerialDate date;
+    private DayDate date;
 
     /** The color for the selected date. */
     private Color dateButtonColor;
@@ -119,7 +119,7 @@ public class SerialDateChooserPanel extends JPanel implements ActionListener {
      */
     public SerialDateChooserPanel() {
 
-        this(SerialDate.createInstance(new Date()), false,
+        this(DayDate.createInstance(new Date()), false,
              DEFAULT_DATE_BUTTON_COLOR,
              DEFAULT_MONTH_BUTTON_COLOR);
 
@@ -132,7 +132,7 @@ public class SerialDateChooserPanel extends JPanel implements ActionListener {
      * @param controlPanel  a flag that indicates whether or not the 'today' button should
      *                      appear on the panel.
      */
-    public SerialDateChooserPanel(final SerialDate date, final boolean controlPanel) {
+    public SerialDateChooserPanel(final DayDate date, final boolean controlPanel) {
 
         this(date, controlPanel,
              DEFAULT_DATE_BUTTON_COLOR,
@@ -148,7 +148,7 @@ public class SerialDateChooserPanel extends JPanel implements ActionListener {
      * @param dateButtonColor  the date button color.
      * @param monthButtonColor  the month button color.
      */
-    public SerialDateChooserPanel(final SerialDate date, final boolean controlPanel,
+    public SerialDateChooserPanel(final DayDate date, final boolean controlPanel,
                                   final Color dateButtonColor, final Color monthButtonColor) {
 
         super(new BorderLayout());
@@ -170,7 +170,7 @@ public class SerialDateChooserPanel extends JPanel implements ActionListener {
      *
      * @param date  the new date.
      */
-    public void setDate(final SerialDate date) {
+    public void setDate(final DayDate date) {
 
         this.date = date;
         this.monthSelector.setSelectedIndex(date.getMonth() - 1);
@@ -184,7 +184,7 @@ public class SerialDateChooserPanel extends JPanel implements ActionListener {
      *
      * @return the selected date.
      */
-    public SerialDate getDate() {
+    public DayDate getDate() {
         return this.date;
     }
 
@@ -197,7 +197,7 @@ public class SerialDateChooserPanel extends JPanel implements ActionListener {
 
         if (e.getActionCommand().equals("monthSelectionChanged")) {
             final JComboBox c = (JComboBox) e.getSource();
-            this.date = SerialDate.createInstance(
+            this.date = DayDate.createInstance(
                 this.date.getDayOfMonth(), c.getSelectedIndex() + 1, this.date.getYYYY()
             );
             refreshButtons();
@@ -206,7 +206,7 @@ public class SerialDateChooserPanel extends JPanel implements ActionListener {
             if (!this.refreshing) {
                 final JComboBox c = (JComboBox) e.getSource();
                 final Integer y = (Integer) c.getSelectedItem();
-                this.date = SerialDate.createInstance(
+                this.date = DayDate.createInstance(
                     this.date.getDayOfMonth(), this.date.getMonth(), y.intValue()
                 );
                 refreshYearSelector();
@@ -214,13 +214,13 @@ public class SerialDateChooserPanel extends JPanel implements ActionListener {
             }
         }
         else if (e.getActionCommand().equals("todayButtonClicked")) {
-            setDate(SerialDate.createInstance(new Date()));
+            setDate(DayDate.createInstance(new Date()));
         }
         else if (e.getActionCommand().equals("dateButtonClicked")) {
             final JButton b = (JButton) e.getSource();
             final int i = Integer.parseInt(b.getName());
-            final SerialDate first = getFirstVisibleDate();
-            final SerialDate selected = SerialDate.addDays(i, first);
+            final DayDate first = getFirstVisibleDate();
+            final DayDate selected = DayDate.addDays(i, first);
             setDate(selected);
         }
 
@@ -266,7 +266,7 @@ public class SerialDateChooserPanel extends JPanel implements ActionListener {
      *
      * @return the button color.
      */
-    protected Color getButtonColor(final SerialDate targetDate) {
+    protected Color getButtonColor(final DayDate targetDate) {
 
         if (this.date.equals(this.date)) {
             return this.dateButtonColor;
@@ -286,12 +286,12 @@ public class SerialDateChooserPanel extends JPanel implements ActionListener {
      *
      * @return the first visible date.
      */
-    protected SerialDate getFirstVisibleDate() {
+    protected DayDate getFirstVisibleDate() {
 
-        SerialDate result = SerialDate.createInstance(1, this.date.getMonth(), this.date.getYYYY());
-        result = SerialDate.addDays(-1, result);
+        DayDate result = DayDate.createInstance(1, this.date.getMonth(), this.date.getYYYY());
+        result = DayDate.addDays(-1, result);
         while (result.getDayOfWeek() != getFirstDayOfWeek()) {
-            result = SerialDate.addDays(-1, result);
+            result = DayDate.addDays(-1, result);
         }
         return result;
 
@@ -311,12 +311,12 @@ public class SerialDateChooserPanel extends JPanel implements ActionListener {
      */
     protected void refreshButtons() {
 
-        SerialDate current = getFirstVisibleDate();
+        DayDate current = getFirstVisibleDate();
         for (int i = 0; i < 42; i++) {
             final JButton button = this.buttons[i];
             button.setText(String.valueOf(current.getDayOfWeek()));
             button.setBackground(getButtonColor(current));
-            current = SerialDate.addDays(1, current);
+            current = DayDate.addDays(1, current);
         }
 
     }
@@ -363,7 +363,7 @@ public class SerialDateChooserPanel extends JPanel implements ActionListener {
      */
     private JPanel constructSelectionPanel() {
         final JPanel p = new JPanel();
-        this.monthSelector = new JComboBox(SerialDate.getMonths());
+        this.monthSelector = new JComboBox(DayDate.getMonths());
         this.monthSelector.addActionListener(this);
         this.monthSelector.setActionCommand("monthSelectionChanged");
         p.add(this.monthSelector);
