+++ src/test/java/org/jfree/date/DatePackageTests.java
@@ -60,7 +60,7 @@ public class DatePackageTests extends TestCase {
      */
     public static Test suite() {
         final TestSuite suite = new TestSuite("org.jfree.date");
-        suite.addTestSuite(SerialDateTest.class);
+        suite.addTestSuite(DayDateTest.class);
         suite.addTestSuite(SerialDateUtilitiesTest.class);
         suite.addTestSuite(SpreadsheetDateTest.class);
         return suite;
