+++ src/test/java/org/jfree/date/SpreadsheetDateTest.java
@@ -61,10 +61,10 @@ import junit.framework.TestSuite;
 public class SpreadsheetDateTest extends TestCase {
 
     /** Date representing 1 January 1900. */
-    private SerialDate jan1Y1900;
+    private DayDate jan1Y1900;
 
     /** Date representing serial day number 2. */
-    private SerialDate s2;
+    private DayDate s2;
 
     /**
      * Returns a test suite for the JUnit test runner.
@@ -98,17 +98,17 @@ public class SpreadsheetDateTest extends TestCase {
      */
     public void test1Jan1900GetDayOfWeek() {
         final int dayOfWeek = this.jan1Y1900.getDayOfWeek();
-        assertEquals(SerialDate.MONDAY, dayOfWeek);
+        assertEquals(DayDate.MONDAY, dayOfWeek);
     }
 
     /**
      * 12 November 2001 is a Monday.
      */
     public void test12Nov2001GetDayOfWeek() {
-        SerialDate nov12Y2001 = new SpreadsheetDate(12, 
+        DayDate nov12Y2001 = new SpreadsheetDate(12,
                 MonthConstants.NOVEMBER, 2001);
         int dayOfWeek = nov12Y2001.getDayOfWeek();
-        assertEquals(SerialDate.MONDAY, dayOfWeek);
+        assertEquals(DayDate.MONDAY, dayOfWeek);
     }
 
     /**
